moved telemetry related code to telemetry file.
