+++ src/client/power-pages/telemetry.ts
@@ -4,6 +4,8 @@
  */
 
 import { ITelemetry } from "../telemetry/ITelemetry";
+import { Tables } from "./create/CreateOperationConstants";
+import * as vscode from "vscode";
 
 // Telemetry Event Names
 export const FileDeleteEvent = 'FileDeleteEvent';
@@ -27,6 +29,11 @@ interface IPowerPagesTelemetryData {
     triggerPoint?: string
 }
 
+export enum TriggerPoint {
+    CONTEXT_MENU = "context-menu",
+    COMMAND_PALETTE = "command-palette",
+}
+
 export function sendTelemetryEvent(telemetry: ITelemetry, telemetryData: IPowerPagesTelemetryData): void {
     const telemetryDataProperties: Record<string, string> = {}
     const telemetryDataMeasurements: Record<string, number> = {}
@@ -47,5 +54,13 @@ export function sendTelemetryEvent(telemetry: ITelemetry, telemetryData: IPowerP
         telemetry.sendTelemetryException(telemetryData.exception, telemetryDataProperties, telemetryDataMeasurements);
     } else {
         telemetry.sendTelemetryEvent(telemetryData.eventName, telemetryDataProperties, telemetryDataMeasurements);
+    }
+}
+
+export function sendTelemetryEventWithTriggerPoint(entityType: Tables, uri:vscode.Uri, telemetry:ITelemetry): void {
+    if (uri) {
+        sendTelemetryEvent(telemetry, { eventName: UserFileCreateEvent, fileEntityType: entityType, triggerPoint: TriggerPoint.CONTEXT_MENU });
+    } else {
+        sendTelemetryEvent(telemetry, { eventName: UserFileCreateEvent, fileEntityType: entityType });
     }
 }
